# services/worker/openapi.yaml
openapi: 3.0.3
info:
  title: Invofox Worker API
  description: |
    Invoice processing worker for the Invofox Invoice Bot.

    This service processes invoices through the pipeline:
    1. Download from Telegram
    2. Upload to Cloud Storage
    3. Extract data with LLM Vision (Gemini/OpenAI)
    4. Append to Google Sheets
    5. Send confirmation to Telegram

    ## Authentication
    - Processing endpoints are called by Cloud Tasks with OIDC tokens
    - Health and metrics endpoints are public

    ## Job States
    | Status | Description |
    |--------|-------------|
    | `pending` | Waiting in queue |
    | `processing` | Currently being processed |
    | `pending_retry` | Failed, waiting for Cloud Tasks retry |
    | `processed` | Successfully completed |
    | `failed` | Failed after max retries |
    | `pending_decision` | Duplicate detected, awaiting user input |
  version: "1.0.0"
  contact:
    name: Invofox Team
servers:
  - url: https://worker-XXXXX.us-central1.run.app
    description: Production
  - url: http://localhost:8081
    description: Local development
...
paths:
...
  /invoice/command:
    post:
      tags:
        - Invoice Generation
      summary: Handle /new command
      description: Starts a new invoice generation session
      operationId: handleNewCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceCommandPayload"
      responses:
        "200":
          description: Command handled
  /invoice/message:
    post:
      tags:
        - Invoice Generation
      summary: Handle invoice conversation message
      description: Processes text messages during invoice generation flow (customer name, amount, description)
      operationId: handleInvoiceMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceMessagePayload"
      responses:
        "200":
          description: Message handled
  /invoice/callback:
    post:
      tags:
        - Invoice Generation
      summary: Handle invoice button callback
      description: |
        Processes inline keyboard callbacks during invoice generation:
        - Document type selection (invoice / invoice-receipt / receipt)
        - Invoice selection (for receipt creation - with pagination support)
        - Show More (pagination - loads next 10 invoices)
        - Payment method selection
        - Confirm / Cancel

        **Receipt Creation Flow:**
        1. User selects "receipt" document type
        2. System shows first 10 open invoices with pagination info
        3. If more than 10 invoices exist, "Show More" button appears
        4. User clicks "Show More" to load next batch (offset-based pagination)
        5. User selects an invoice to create receipt for

        **Pagination:**
        - Default page size: 10 invoices
        - Show More button displays current position (e.g., "10/45")
        - Message updates to show range (e.g., "מציג 11-20 מתוך 45 חשבוניות")

        **Callback Actions:**
        - `select_type`: Choose document type (invoice/invoice_receipt/receipt)
        - `select_invoice`: Select invoice for receipt creation
        - `show_more`: Load next page of invoices (includes offset)
        - `select_payment`: Choose payment method
        - `confirm`: Confirm and generate document
        - `cancel`: Cancel operation
      operationId: handleInvoiceCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceCallbackPayload"
            examples:
              selectType:
                summary: Document type selection
                value:
                  type: callback
                  chatId: -123456
                  userId: 789
                  messageId: 100
                  username: testuser
                  callbackQueryId: cb123
                  data: '{"action":"select_type","documentType":"receipt"}'
              selectInvoice:
                summary: Invoice selection for receipt
                value:
                  type: callback
                  chatId: -123456
                  userId: 789
                  messageId: 101
                  username: testuser
                  callbackQueryId: cb124
                  data: '{"action":"select_invoice","invoiceNumber":"I-2026-15"}'
              showMore:
                summary: Pagination - load next 10 invoices
                value:
                  type: callback
                  chatId: -123456
                  userId: 789
                  messageId: 101
                  username: testuser
                  callbackQueryId: cb125
                  data: '{"action":"show_more","offset":10}'
              selectPayment:
                summary: Payment method selection
                value:
                  type: callback
                  chatId: -123456
                  userId: 789
                  messageId: 102
                  username: testuser
                  callbackQueryId: cb126
                  data: '{"action":"select_payment","paymentMethod":"מזומן"}'
              confirm:
                summary: Confirm and generate document
                value:
                  type: callback
                  chatId: -123456
                  userId: 789
                  messageId: 103
                  username: testuser
                  callbackQueryId: cb127
                  data: '{"action":"confirm"}'
      responses:
        "200":
          description: Callback handled
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  action:
                    type: string
                    enum: [type_selected, showing_open_invoices, invoice_selected, showed_more_invoices, no_more_invoices, payment_selected, invoice_generated, cancelled]
              examples:
                showingInvoices:
                  summary: Showed open invoices with pagination
                  value:
                    ok: true
                    action: showing_open_invoices
                showedMoreInvoices:
                  summary: Loaded next batch of invoices
                  value:
                    ok: true
                    action: showed_more_invoices
                noMoreInvoices:
                  summary: No more invoices to show
                  value:
                    ok: true
                    action: no_more_invoices
...
components:
  securitySchemes:
    cloudTasks:
      type: http
      scheme: bearer
      bearerFormat: OIDC
      description: Cloud Tasks OIDC token
  schemas:
...
    InvoiceCommandPayload:
      type: object
      properties:
        type:
          type: string
          enum: [command]
        chatId:
          type: integer
        messageId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        firstName:
          type: string
        text:
          type: string
          description: Full command text
        receivedAt:
          type: string
          format: date-time
    InvoiceMessagePayload:
      type: object
      properties:
        type:
          type: string
          enum: [message]
        chatId:
          type: integer
        messageId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        firstName:
          type: string
        text:
          type: string
        receivedAt:
          type: string
          format: date-time
    InvoiceCallbackPayload:
      type: object
      required:
        - type
        - callbackQueryId
        - chatId
        - messageId
        - userId
        - data
      properties:
        type:
          type: string
          enum: [callback]
          description: Payload type identifier
        callbackQueryId:
          type: string
          description: Telegram callback query ID (for answering)
        chatId:
          type: integer
          description: Telegram chat ID
        messageId:
          type: integer
          description: Message ID containing the inline keyboard
        userId:
          type: integer
          description: User ID who clicked the button
        username:
          type: string
          description: Username (optional)
        data:
          type: string
          description: |
            JSON-stringified callback action. Structure varies by action type:

            **select_type**: `{"action":"select_type","documentType":"invoice"|"invoice_receipt"|"receipt"}`

            **select_invoice**: `{"action":"select_invoice","invoiceNumber":"I-2026-15"}`

            **show_more**: `{"action":"show_more","offset":10}` (offset for pagination)

            **select_payment**: `{"action":"select_payment","paymentMethod":"מזומן"|"ביט"|"PayBox"|"העברה"|"אשראי"|"צ'ק"}`

            **confirm**: `{"action":"confirm"}`

            **cancel**: `{"action":"cancel"}`
          example: '{"action":"select_type","documentType":"receipt"}'
...